pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "cedar-detect"
description = ""
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Steven Rosenthal", email = "" },
    { name = "Leo Covarrubias", email = "leo@leocov.com" },
]
requires-python = ">=3.7"
keywords = ["astronomy"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dynamic = ["version"]

dependencies = [
    "numpy >= 1.21.1",
    "Pillow >= 8.3.1",
]

# -----------------------------------------------------------------------------

[tool.hatch.version]
path = "python/cedar_detect/__init__.py"

# -----------------------------------------------------------------------------
# dev environment -------------------------------------------------------------
[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
dev-mode = true
extra-dependencies = [
    "maturin>=1.0,<2.0",
    "pytest",
]

[tool.hatch.envs.default.scripts]
setup = "maturin develop"

# -----------------------------------------------------------------------------
# maturin config --------------------------------------------------------------
[tool.maturin]
bindings = "pyo3"
python-source = "python"
module-name = "cedar_detect._cedar_detect"
features = ["pyo3/extension-module"]
sdist = false

[tool.pytest.ini_options]
pythonpath = "./python"
testpaths = [
    "tests/unit/*",
]
